pyproject.toml
[project]
name = "fhir-structuring-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "click>=8.3.1",
    "fhir-resources>=8.2.0",
    "httpx>=0.28.1",
    "logfire>=4.30.0",
    "mcp>=1.26.0",
    "pydantic-ai>=1.70.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "bandit>=1.8.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "types-requests>=2.32.0",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# Exclude tests and scripts from strict checking (common practice)
[[tool.mypy.overrides]]
module = ["tests.*", "scripts.*"]
disallow_untyped_defs = false
disallow_untyped_calls = false
disallow_incomplete_defs = false
check_untyped_defs = false

# Exclude MCP server, CLI, and agents due to untyped decorators from third-party libraries
[[tool.mypy.overrides]]
module = ["src.fhir_doc_tool.server", "src.fhir_doc_tool.cli", "src.main", "src.clinical_analyst.agent", "src.validator.agent", "src.clinical_analyst.mcp_client"]
disallow_untyped_decorators = false
disallow_untyped_calls = false

[tool.bandit]
exclude_dirs = ["tests", "scripts"]
skips = ["B101"]  # Skip assert_used (common in tests despite exclusion)
